authenticate service requests
In order to interact with the API, the requests must be authenticated
with an API token. The access token is currently set via environment
variable. Ideally, this would never be handled on the client side of
things and would be done through an API that provides authenticated
sessions to the client.